pyproject.toml
[project]
name = "s3"
version = "0.1.0"
requires-python = ">=3.11.11"
dependencies = [
    "aws-cdk-lib==2.202.0",
    "constructs>=10.0.0,<11.0.0",
]

[dependency-groups]
dev = [
    "pytest==6.2.5",
]
